feat: initial support to boot the HPE GSC ASIC
This commit supports booting a limited OpenBMC on the HPE GSC ASIC, a custom
BMC ASIC found in all HPE Proliant Gen12 Servers. Support for many OpenBMC
functions, such as IPMI, Redfish & firmware update are functional, but support
for booting the host itself is not ready yet and was not included in this
commit.